[REFACTOR] Auth 문서 테스트를 API별 클래스로 분리
