+++ src/test/java/org/websoso/WSSServer/auth/controller/auth/PostAuthLogoutDocsTest.java
@@ -1,4 +1,4 @@
-package org.websoso.WSSServer.auth.controller;
+package org.websoso.WSSServer.auth.controller.auth;
 
 import static com.epages.restdocs.apispec.ResourceDocumentation.resource;
 import static org.mockito.ArgumentMatchers.any;
@@ -45,6 +45,7 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
 import org.websoso.WSSServer.application.AccountApplication;
 import org.websoso.WSSServer.auth.application.AuthApplication;
+import org.websoso.WSSServer.auth.controller.AuthController;
 import org.websoso.WSSServer.auth.controller.dto.LogoutRequest;
 import org.websoso.WSSServer.auth.service.AppleService;
 import org.websoso.WSSServer.exception.exception.CustomKakaoException;
@@ -61,7 +62,7 @@ import org.websoso.WSSServer.user.service.UserService;
  */
 @AutoConfigureRestDocs
 @AuthenticatedControllerTest(AuthController.class)
-class AuthControllerDocsTest {
+class PostAuthLogoutDocsTest {
 
     private static final Long USER_ID = 42L;
     private static final LogoutRequest LOGOUT_REQUEST = new LogoutRequest("refresh-token", "device-identifier");
